backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
sqlparse==0.5.3
